integrated current_user

+++ routers/members.py
@@ -11,10 +11,14 @@ router = APIRouter()
 
 @router.post("/members")
 def create_member_endpoint(member_data , db=Depends(get_db), current_user: User = Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     return create_member(db , member_data)
 
 @router.get("/members/{member_id}")
 def get_member_endpoint(member_id : int , db=Depends(get_db), current_user: User = Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     member = get_member(db , member_id)
     if not member :
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Member not found")
@@ -22,12 +26,18 @@ def get_member_endpoint(member_id : int , db=Depends(get_db), current_user: User
 
 @router.get("/members")
 def get_all_members_endpoint(db=Depends(get_db), current_user: User = Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     return get_all_members(db)
 
 @router.put("/members/{member_id}")
 def update_member_endpoint(member_id : int , new_data : dict , db=Depends(get_db), current_user: User = Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     return update_member(db , member_id , new_data)
 
 @router.delete("/members/{member_id}")
 def delete_member_endpoint(member_id : int , db=Depends(get_db), current_user: User = Depends(get_current_user)):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
     return delete_member(db , member_id)
